ai-service/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
pydantic
pydantic-settings
python-dotenv
python-multipart
redis
httpx
jinja2
loguru
